Fix: 판매자가 아닐 경우에도 수정, 삭제가 되는 경우 없앰
- Goods와 Member를 seller_id로 join 해놓음.
- jwt 생성시 claims를 member.id까지 저장해놓음

+++ src/main/java/com/example/hs/domain/goods/service/GoodsService.java
@@ -6,7 +6,11 @@ import static com.example.hs.global.exception.ErrorCode.CHECK_IF_SOLD_OUT_GOODS;
 import static com.example.hs.global.exception.ErrorCode.INVALID_GOODS_STATUS_AT_FIRST;
 import static com.example.hs.global.exception.ErrorCode.INVALID_MIN_QUANTITY_AT_FIRST;
 import static com.example.hs.global.exception.ErrorCode.NOT_FOUND_GOODS;
+import static com.example.hs.global.exception.ErrorCode.NOT_FOUND_MEMBER;
+import static com.example.hs.global.exception.ErrorCode.NOT_MATCH_SELLER;
 
+import com.example.hs.domain.auth.entity.Member;
+import com.example.hs.domain.auth.repository.MemberRepository;
 import com.example.hs.domain.goods.dto.GoodsDto;
 import com.example.hs.domain.goods.dto.GoodsEditRequest;
 import com.example.hs.domain.goods.dto.GoodsRequest;
@@ -14,8 +18,11 @@ import com.example.hs.domain.goods.entity.Goods;
 import com.example.hs.domain.goods.repository.GoodsRepository;
 import com.example.hs.domain.goods.type.GoodsStatus;
 import com.example.hs.global.exception.CustomException;
+import com.example.hs.global.security.userdetails.CustomUserDetails;
 import java.util.List;
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -23,6 +30,7 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class GoodsService {
   private final GoodsRepository goodsRepository;
+  private final MemberRepository memberRepository;
 
   public List<GoodsDto> getAll() {
     return GoodsDto.fromEntity(goodsRepository.findAll());
@@ -34,7 +42,7 @@ public class GoodsService {
   }
 
   @Transactional
-  public GoodsDto postGoods(GoodsRequest goodsRequest) {
+  public GoodsDto postGoods(GoodsRequest goodsRequest, CustomUserDetails member) {
     if (goodsRequest.getQuantity() == 0) {
       throw new CustomException(INVALID_MIN_QUANTITY_AT_FIRST);
     }
@@ -43,21 +51,30 @@ public class GoodsService {
       throw new CustomException(INVALID_GOODS_STATUS_AT_FIRST);
     }
 
+    Member seller = memberRepository.findById(member.getId())
+        .orElseThrow(() -> new CustomException(NOT_FOUND_MEMBER));
+
     Goods goods = goodsRepository.save(Goods.builder()
         .goodsName(goodsRequest.getGoodsName())
         .description(goodsRequest.getDescription())
         .price(goodsRequest.getPrice())
         .quantity(goodsRequest.getQuantity())
         .goodsStatus(goodsRequest.getGoodsStatus())
+        .seller(seller)
         .build());
 
     return GoodsDto.fromEntity(goods);
   }
 
   @Transactional
-  public GoodsDto updateGoods(long goodsId, GoodsEditRequest goodsEditRequest) {
+  public GoodsDto updateGoods(long goodsId,
+      GoodsEditRequest goodsEditRequest, CustomUserDetails member) {
     Goods goods = goodsRepository.findById(goodsId)
         .orElseThrow(() -> new CustomException(NOT_FOUND_GOODS));
+
+    if (member.getId() != goods.getSeller().getId()) {
+      throw new CustomException(NOT_MATCH_SELLER);
+    }
 
     if ((goodsEditRequest.getQuantity() == 0 && goodsEditRequest.getGoodsStatus() != SOLD_OUT)
         || (goodsEditRequest.getGoodsStatus() == SOLD_OUT && goodsEditRequest.getQuantity() != 0)) {
@@ -74,9 +91,14 @@ public class GoodsService {
   }
 
   @Transactional
-  public String deleteGoods(long goodsId) {
+  public String deleteGoods(long goodsId, CustomUserDetails member) {
     Goods goods = goodsRepository.findById(goodsId)
         .orElseThrow(() -> new CustomException(NOT_FOUND_GOODS));
+
+    if (member.getId() != goods.getSeller().getId()) {
+      throw new CustomException(NOT_MATCH_SELLER);
+    }
+
     goodsRepository.delete(goods);
     return String.format("%s 상품이 삭제되었습니다.", goods.getGoodsName());
   }
